+++ src/arch/cc/vcd_cc.h
@@ -21,7 +21,7 @@ public:     // funcs
     vcd_cc() = default;
     ~vcd_cc() = default;
 
-    void programStart(int qubitNumber, int cycleTime, int maxGroups, const settings_cc &settings);
+    void programStart(size_t qubitNumber, int cycleTime, int maxGroups, const settings_cc &settings);
     void programFinish(const std::string &progName);
     void kernelFinish(const std::string &kernelName, size_t durationInCycles);
     void bundleFinishGroup(size_t startCycle, unsigned int durationInCycles, uint32_t groupDigOut, const std::string &signalValue, int instrIdx, int group);
